funding-rollover-bot/requirements.txt
aiohttp>=3.9.0
websockets>=12.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
pyyaml>=6.0
structlog>=23.0.0
scipy>=1.11.0
aiosqlite>=0.19.0
python-dateutil>=2.8.0
ntplib>=0.4.0
